Add files via upload
Given a string, find a substring based on the following conditions: 

The substring must be the longest one of all the possible substring in the given string. 
There must not be any repeating characters in the substring. 
If there is more than one substring satisfying the above two conditions, then print the substring which occurs first. 
If there is no substring satisfying all the aforementioned conditions then print -1.